feat!: Ability to generate changelog from previous tag ..
.. even if `automatic_release_tag` is given. Previous behaviour would always
try to generate changelog from the same tag, seemingly assuming that you use
the same one every time -- frequently resulting in an empty changelog.

Added a parameter `is_tag_static` to allow the previous behaviour when set to
`true`, but it defaults to `false` (as this is what I believe to be the sane
behaviour), so this is a breaking change.

BREAKING CHANGE: Changelog now generates from previous tag instead of same tag
by default even when `automatic_release_tag` is given.